+++ bgstally/windows/colonisation.py
@@ -154,6 +154,7 @@ class ColonisationWindow:
                             }
         # UI components'
         self.window:tk.Toplevel|None = None
+        self.tl:dict = {} # Map of tab number to system index
         self.tabbar:ScrollableNotebook|None = None
         self.add_dialog:tk.Frame|None = None
         self.react:tk.Frame|None = None
@@ -201,16 +202,19 @@ class ColonisationWindow:
             Debug.logger.info(f"No systems so not creating colonisation section")
             return
 
+        tabnum:int = 1
         for sysnum, system in enumerate(systems): # Create a frame for the sytem
-            tabnum = sysnum + 1
+            if system.get('Hidden', False) == True: continue
+            self.tl[tabnum] = sysnum
             self._create_system_tab(tabnum, system)
+            tabnum += 1
 
         if tabnum > 0:
-            t = 0
-            for t in range(0, tabnum-1):
-                if systems[t].get('Hidden', True) == False:
+            for t in range(1, tabnum-1):
+                Debug.logger.debug(f"{t} {self.tl.get(t)} {tabnum-1}")
+                if systems[self.tl[t]].get('Hidden', True) == False:
                     break
-            self.tabbar.select(t+1) # Select the first non-hidden system tab
+            self.tabbar.select(t) # Select the first non-hidden system tab
         self.tabbar.pack(fill=tk.BOTH, side=tk.TOP, expand=True, padx=5, pady=5)
 
 
@@ -255,7 +259,7 @@ class ColonisationWindow:
     @catch_exceptions
     def _create_title_frame(self, tabnum:int, tab:ttk.Frame) -> None:
         ''' Create the title frame with system name and tick info '''
-        sysnum:int = tabnum -1
+        sysnum:int = self.tl[tabnum]
         systems:list = self.colonisation.get_all_systems()
 
         title_frame:ttk.Frame = ttk.Frame(tab, style="Title.TFrame")
@@ -266,7 +270,7 @@ class ColonisationWindow:
         style.configure("Title.TFrame")
 
         # System name label
-        while len(self.plan_titles) <= sysnum:
+        while len(self.plan_titles) <= tabnum:
             self.plan_titles.append({})
 
 
@@ -279,7 +283,7 @@ class ColonisationWindow:
         else:
             name_label = ttk.Label(title_frame, text=name, font=FONT_HEADING_1, foreground=COLOUR_HEADING_1)
         name_label.pack(side=tk.LEFT, padx=10, pady=5)
-        self.plan_titles[sysnum]['Name'] = name_label
+        self.plan_titles[tabnum]['Name'] = name_label
 
         if sysname != '':
             sys_label:ttk.Label = ttk.Label(title_frame, text=sysname, cursor="hand2")
@@ -289,7 +293,7 @@ class ColonisationWindow:
             sys_label.bind("<Button-3>", partial(self._context_menu, systems[sysnum], 'System'))
             ToolTip(sys_label, text=_("Left click view system, right click menu")) # LANG: tooltip for the copy to clipboard icon
             self._set_weight(sys_label)
-            self.plan_titles[sysnum]['System'] = sys_label
+            self.plan_titles[tabnum]['System'] = sys_label
 
         if systems[sysnum].get('Bodies', None) != None and len(systems[sysnum]['Bodies']) > 0:
             bodies:str = str(len(systems[sysnum]['Bodies'])) + " " + _("Bodies") # LANG: bodies in the system
@@ -465,7 +469,7 @@ class ColonisationWindow:
         text:tk.Text = tk.Text(self.bodies_fr, font=FONT_SMALL, yscrollcommand=scr.set)
         text.pack(fill=tk.BOTH, side=tk.TOP, expand=True, padx=5, pady=5)
 
-        sysnum:int = tabnum - 1
+        sysnum:int = self.tl[tabnum]
         systems:list = self.colonisation.get_all_systems()
 
         bodies:list = systems[sysnum].get('Bodies', None)
@@ -524,7 +528,7 @@ class ColonisationWindow:
             self.notes_fr.destroy()
             self.notes_fr
 
-        sysnum:int = tabnum -1
+        sysnum:int = self.tl[tabnum]
         systems:list = self.colonisation.get_all_systems()
 
         if self.notes_fr != None and self.notes_fr.winfo_exists():
@@ -594,21 +598,24 @@ class ColonisationWindow:
         sheet.set_sheet_data(data)
 
         self._config_sheet(sheet, system)
-        sheet.enable_bindings('single_select', 'drag_select', 'edit_cell', 'arrowkeys', 'right_click_popup_menu', 'copy', 'cut', 'paste', 'delete', 'undo')
-        sheet.edit_validation(func=partial(self._validate_edits, sheet))
-        sheet.extra_bindings(['all_modified_events', 'cell_select', 'ctrl_row_select', 'rc_delete_row', 'rc_insert_row'], func=partial(self.sheet_modified, sheet, tabnum))
-
-        sheet.popup_menu_add_command(label="Move up", func=partial(self._row_modified, sheet, tabnum, 'MoveUp'), image=tk.PhotoImage(data=ICON_SORT_DESC), compound="left")
-        sheet.popup_menu_add_command(label="Move down", func=partial(self._row_modified, sheet, tabnum, 'MoveDown'), image=tk.PhotoImage(data=ICON_SORT_ASC), compound="left")
-        sheet.popup_menu_add_command(label="Insert build above", func=partial(self._row_modified, sheet, tabnum, 'InsAbove'), image=tk.PhotoImage(data=ICON_ADD), compound="left")
-        sheet.popup_menu_add_command(label="Insert build below", func=partial(self._row_modified, sheet, tabnum, 'InsBelow'), image=tk.PhotoImage(data=ICON_ADD), compound="left")
-        sheet.popup_menu_add_command(label="Delete build", func=partial(self._row_modified, sheet, tabnum, 'Delete'), image=tk.PhotoImage(data=ICON_DEL), compound="left")
-        sheet.popup_menu_add_command(label="Toggle Readonly", func=partial(self._row_modified, sheet, tabnum, 'Readonly'), image=tk.PhotoImage(data=ICON_REDO), compound="left")
-
-        if len(self.sheets) < tabnum:
-            self.sheets.append(sheet)
+        if system.get('RCSync', False) == True and RavenColonial(self.colonisation).is_editable(system) == False:
+            sheet.enable_bindings('single_select', 'drag_select', 'arrowkeys')
         else:
-            self.sheets[tabnum-1] = sheet
+            sheet.enable_bindings('single_select', 'drag_select', 'edit_cell', 'arrowkeys', 'right_click_popup_menu', 'copy', 'cut', 'paste', 'delete', 'undo')
+
+            sheet.edit_validation(func=partial(self._validate_edits, sheet))
+            sheet.extra_bindings(['all_modified_events', 'cell_select', 'ctrl_row_select', 'rc_delete_row', 'rc_insert_row'], func=partial(self.sheet_modified, sheet, tabnum))
+
+            sheet.popup_menu_add_command(label="Move up", func=partial(self._row_modified, sheet, tabnum, 'MoveUp'), image=tk.PhotoImage(data=ICON_SORT_DESC), compound="left")
+            sheet.popup_menu_add_command(label="Move down", func=partial(self._row_modified, sheet, tabnum, 'MoveDown'), image=tk.PhotoImage(data=ICON_SORT_ASC), compound="left")
+            sheet.popup_menu_add_command(label="Insert build above", func=partial(self._row_modified, sheet, tabnum, 'InsAbove'), image=tk.PhotoImage(data=ICON_ADD), compound="left")
+            sheet.popup_menu_add_command(label="Insert build below", func=partial(self._row_modified, sheet, tabnum, 'InsBelow'), image=tk.PhotoImage(data=ICON_ADD), compound="left")
+            sheet.popup_menu_add_command(label="Delete build", func=partial(self._row_modified, sheet, tabnum, 'Delete'), image=tk.PhotoImage(data=ICON_DEL), compound="left")
+            sheet.popup_menu_add_command(label="Toggle Readonly", func=partial(self._row_modified, sheet, tabnum, 'Readonly'), image=tk.PhotoImage(data=ICON_REDO), compound="left")
+
+        while len(self.sheets) <= tabnum:
+            self.sheets.append(sheet)
+        self.sheets[tabnum] = sheet
 
 
     @catch_exceptions
@@ -621,7 +628,7 @@ class ColonisationWindow:
         row:int = selected.row - FIRST_BUILD_ROW                            # type: ignore
         if row < 0: return
 
-        sysnum:int = tabnum -1
+        sysnum:int = self.tl[tabnum]
         systems:list = self.colonisation.get_all_systems()
         system:dict = systems[sysnum]
 
@@ -637,7 +644,7 @@ class ColonisationWindow:
             case 'Delete':
                 self.colonisation.remove_build(system, row)
                 sheet.del_row(row)
-                self._config_sheet(self.sheets[sysnum], system)
+                self._config_sheet(self.sheets[tabnum], system)
             case 'MoveUp':
                 if row == 0:
                     return
@@ -833,20 +840,16 @@ class ColonisationWindow:
         scol:int = 0
         new:list = self._build_summary(system)
 
-        Debug.logger.debug(f"System: {system.get('Name', '')}")
-        Debug.logger.debug(f"System: {new}")
         for i, x in enumerate(self.summary_rows.keys()):
             for j, details in enumerate(self.summary_cols.values()):
                 sheet[self._cell(i+srow,j)].data = ' ' if new[i][j] == 0 else f"{new[i][j]:,}" if details.get('format') == 'int' else new[i][j]
                 if details.get('background') != None:
                     sheet[self._cell(i+srow,j+scol)].highlight(bg=self._set_background(details.get('background'), new[i][j], details.get('max', 1)))
 
-        Debug.logger.debug(f"Security state: {system.get('Name', '')} {system.get('RCSync', False)}")
         if system.get('RCSync', False) == True:
             sheet[self._cell(srow,list(self.summary_cols.keys()).index('RCOpen'))].dropdown(values=['Open', 'Secured'])
             sheet[self._cell(srow,list(self.summary_cols.keys()).index('RCOpen'))].data = 'Open' if system.get('RCOpen', True) == True else 'Secured'
         else:
-            Debug.logger.debug(f"Setting readonly security")
             sheet[self._cell(srow,list(self.summary_cols.keys()).index('RCOpen'))].del_dropdown()
             sheet[self._cell(srow,list(self.summary_cols.keys()).index('RCOpen'))].readonly()
             sheet[self._cell(srow,list(self.summary_cols.keys()).index('RCOpen'))].data = "N/A"
@@ -1052,14 +1055,15 @@ class ColonisationWindow:
         if self.window == None: return
 
         systems:list = self.colonisation.get_all_systems()
-        for i, tab in enumerate(self.sheets):
-            system:dict = systems[i]
-            self._update_title(i, system)
-            self._update_summary(FIRST_SUMMARY_ROW, self.sheets[i], system)
-            self._update_detail(FIRST_BUILD_ROW, self.sheets[i], system)
+        for tabnum, sysnum in self.tl.items():
+            system:dict = systems[sysnum]
+            #Debug.logger.debug(f"System: t {tabnum} sys {sysnum} {system.get('Name', '')}")
+            self._update_title(tabnum, system)
+            self._update_summary(FIRST_SUMMARY_ROW, self.sheets[tabnum], system)
+            self._update_detail(FIRST_BUILD_ROW, self.sheets[tabnum], system)
             # Not our system? Then it's readonly
             if system.get('RCSync', False) == True and RavenColonial(self.colonisation).is_editable(system) == False:
-                self.sheets[i]['B1:Z'].readonly()
+                self.sheets[tabnum]['B1:Z'].readonly()
 
 
     @catch_exceptions
@@ -1080,7 +1084,7 @@ class ColonisationWindow:
     def sheet_modified(self, sheet:Sheet, tabnum:int, event = None) -> None:
         ''' Handle edits to the sheet. This is where we update the system data. '''
 
-        sysnum:int = tabnum -1
+        sysnum:int = self.tl[tabnum]
         systems:list = self.colonisation.get_all_systems()
         system:dict = systems[sysnum]
 
@@ -1137,10 +1141,10 @@ class ColonisationWindow:
                 else:
                     self.colonisation.set_base_type(system, row, val)
 
-                sdata:list = self.sheets[sysnum].data
+                sdata:list = self.sheets[tabnum].data
                 sdata.pop(row + FIRST_BUILD_ROW)
-                self.sheets[sysnum].set_sheet_data(sdata, redraw=False)
-                self._config_sheet(self.sheets[sysnum], system)
+                self.sheets[tabnum].set_sheet_data(sdata, redraw=False)
+                self._config_sheet(self.sheets[tabnum], system)
 
             case 'Base Type' | 'Layout' if val != ' ':
                 self.colonisation.set_base_type(system, row, val)
@@ -1153,8 +1157,8 @@ class ColonisationWindow:
                 sdata.append(self._get_detail_header())
                 sdata += self._build_detail(system)
 
-                self.sheets[sysnum].set_sheet_data(sdata, redraw=False)
-                self._config_sheet(self.sheets[sysnum], system)
+                self.sheets[tabnum].set_sheet_data(sdata, redraw=False)
+                self._config_sheet(self.sheets[tabnum], system)
 
             case 'Body':
                 # If the body is set, update the build data
@@ -1268,7 +1272,8 @@ class ColonisationWindow:
 
         # Destroy the existing frames and recreate them.
         self.tabbar.destroy()
-        for s in self.sheets: s.destroy()
+        for s in self.sheets:
+            s.destroy()
         self.sheets = []
         self._create_frames()   # Create main frames
         self.update_display()   # Populate them
@@ -1289,7 +1294,8 @@ class ColonisationWindow:
             return
 
         systems:list = self.colonisation.get_all_systems()
-        self._create_system_tab(len(systems), system)
+        tabnum:int = max(self.tl.keys())+1 # Next available tab
+        self._create_system_tab(tabnum, system)
         self.update_display()
 
 
@@ -1297,7 +1303,7 @@ class ColonisationWindow:
     def edit_system_dialog(self, tabnum:int, btn:ttk.Button) -> None:
         ''' Show dialog to edit a system '''
 
-        sysnum:int = tabnum -1
+        sysnum:int = self.tl[tabnum]
         systems:list = self.colonisation.get_all_systems()
         if sysnum > len(systems):
             Debug.logger.info(f"Invalid tab {tabnum} {sysnum}")
@@ -1368,7 +1374,7 @@ class ColonisationWindow:
     @catch_exceptions
     def _edit_system(self, tabnum:int, name:str, sysname:str, rcsync:bool, hide:bool, dialog:tk.Toplevel) -> None:
         ''' Edit a system's plan, name and sync state '''
-        sysnum:int = tabnum -1
+        sysnum:int = self.tl[tabnum]
 
         if not name:
             messagebox.showerror(_("Error"), _("Plan name is required"))
@@ -1394,7 +1400,7 @@ class ColonisationWindow:
     @catch_exceptions
     def delete_system(self, tabnum:int, tab: ttk.Frame) -> None:
         ''' Remove the current system '''
-        sysnum:int = tabnum -1
+        sysnum:int = self.tl[tabnum]
         # Confirm removal
         if not messagebox.askyesno(
             _("Confirm Removal"),
@@ -1408,8 +1414,8 @@ class ColonisationWindow:
         Debug.logger.info(f"Deleting system {tabnum}")
         tabs:list = self.tabbar.tabs()
         self.tabbar.forget(tabs[tabnum])
-        del self.sheets[sysnum]
-        del self.plan_titles[sysnum]
+        del self.sheets[tabnum]
+        del self.plan_titles[tabnum]
         self.colonisation.remove_system(sysnum)
 
         self.update_display()
@@ -1419,7 +1425,7 @@ class ColonisationWindow:
     def _rc_refresh_system(self, tabnum:int) -> None:
         ''' Reload the current system from RavenColonial '''
 
-        sysnum:int = tabnum -1
+        sysnum:int = self.tl[tabnum]
         systems:list = self.colonisation.get_all_systems()
         if sysnum > len(systems):
             Debug.logger.info(f"Invalid tab {tabnum} {sysnum}")